Refactor out OpenXRHandJointProvider
