+++ Assets/MRTK/Providers/OpenXR/Scripts/OpenXRHandJointProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f7b7c3cede388f54eaeac3b3b0d80f83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
